external/RAE/requirements.txt
timm==0.9.16
accelerate==0.23.0
torchdiffeq==0.2.5
wandb
transformers==4.56.2
einops
omegaconf==2.3.0
numpy==1.26.0
torchmetrics
torch-fidelity